feat: complete entitys crud

+++ src/main/java/com/products/products/DTOs/ProductMaterialResponseDTO.java
@@ -3,11 +3,13 @@ package com.products.products.DTOs;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class ProductMaterialResponseDTO {
     private String id;
     private String materialId;
     private String materialName;
-    private Integer quantityRequired;
+    private BigDecimal quantityRequired;
 }
